+++ backend/src/__tests__/controllers/locationController.test.ts
@@ -1,11 +1,11 @@
 import { Request, Response } from "express";
 import { LocationService } from "@/modules/locations/services/locationService";
-import { sendResponse } from "@/utils/response";
+import { sendResponse } from "@/modules/infrastructure/utilities/response";
 import { APIResponse } from "@/types/response";
 import { getAllLocations } from "@/modules/locations/controllers/locationController";
 
 jest.mock("@/modules/locations/services/locationService");
-jest.mock("@/utils/response");
+jest.mock("@/modules/infrastructure/utilities/response");
 
 describe("Location Controller", () => {
   let mockRequest: Partial<Request>;
