+++ backend/src/__tests__/middleware/middleware.test.ts
@@ -2,10 +2,10 @@ import request from "supertest";
 import express from "express";
 import supabase from "@/modules/shared/services/supabaseClient";
 import { serverMiddleare, verifyAndGetUser } from "@/middleware/middleware";
-import { sendResponse } from "@/utils/response";
+import { sendResponse } from "@/modules/infrastructure/utilities/response";
 
 jest.mock("@/modules/shared/services/supabaseClient");
-jest.mock("@/utils/response");
+jest.mock("@/modules/infrastructure/utilities/response");
 
 const app = express();
 app.use(express.json());
